src/requirements.txt
Django==2.1.2
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
social-auth-app-django==2.0.0
social-auth-core==1.5.0
google-api-python-client==1.7.4
python-dateutil==2.7.5
httplib2
oauth2client
django-background-tasks==1.2.0
beautifulsoup4==4.6.3
